fix: 불필요한 코드 삭제

+++ src/test/java/nextstep/users/infrastructure/UserRepositoryTest.java
@@ -31,7 +31,7 @@ public class UserRepositoryTest {
 
     @Test
     public void save(){
-        Long saved = userRepository.save(new NsUser(null,"test", "test", "test", "test@test.net"));
+        int saved = userRepository.save(new NsUser(null,"test", "test", "test", "test@test.net"));
         System.out.println(saved);
     }
